offset_for_local_datetime: Add 'fall_forward_on_invalid_local_time' arg
If this is true and the local time for the given $dt object does
not exist in the time zone (due to DST changes for example), the previous
span will be returned.

This way you can be sure to get a value rather than an exception, if you
really want.